findByIds

+++ codegen/__out__/pc/src/views/nuxt/seo/Api.ts
@@ -222,9 +222,12 @@ export async function updateById(
  * 根据 id 查找SEO优化
  */
 export async function findById(
-  id: SeoId,
+  id?: SeoId,
   opt?: GqlOpt,
-) {
+): Promise<SeoModel | undefined> {
+  if (!id) {
+    return;
+  }
   const data: {
     findByIdSeo?: SeoModel;
   } = await query({
@@ -245,12 +248,52 @@ export async function findById(
 }
 
 /**
+ * 根据 ids 查找SEO优化
+ */
+export async function findByIds(
+  ids: SeoId[],
+  opt?: GqlOpt,
+): Promise<SeoModel[]> {
+  if (ids.length === 0) {
+    return [ ];
+  }
+  opt = opt || { };
+  opt.showErrMsg = false;
+  let models: SeoModel[] = [ ];
+  try {
+    const data: {
+      findByIdsSeo: SeoModel[];
+    } = await query({
+      query: `
+        query($ids: [SeoId!]!) {
+          findByIdsSeo(ids: $ids) {
+            ${ seoQueryField }
+          }
+        }
+      `,
+      variables: {
+        ids,
+      },
+    }, opt);
+    models = data.findByIdsSeo;
+  } catch (_err) { /* empty */ }
+  for (let i = 0; i < models.length; i++) {
+    const model = models[i];
+    await setLblById(model);
+  }
+  return models;
+}
+
+/**
  * 根据 ids 删除SEO优化
  */
 export async function deleteByIds(
   ids: SeoId[],
   opt?: GqlOpt,
-) {
+): Promise<number> {
+  if (ids.length === 0) {
+    return 0;
+  }
   const data: {
     deleteByIdsSeo: Mutation["deleteByIdsSeo"];
   } = await mutation({
@@ -271,9 +314,12 @@ export async function deleteByIds(
  * 根据 id 设置默认SEO优化
  */
 export async function defaultById(
-  id: SeoId,
+  id?: SeoId,
   opt?: GqlOpt,
 ) {
+  if (!id) {
+    return 0;
+  }
   const data: {
     defaultByIdSeo: Mutation["defaultByIdSeo"];
   } = await mutation({
@@ -297,7 +343,10 @@ export async function lockByIds(
   ids: SeoId[],
   is_locked: 0 | 1,
   opt?: GqlOpt,
-) {
+): Promise<number> {
+  if (ids.length === 0) {
+    return 0;
+  }
   const data: {
     lockByIdsSeo: Mutation["lockByIdsSeo"];
   } = await mutation({
@@ -321,7 +370,10 @@ export async function lockByIds(
 export async function revertByIds(
   ids: SeoId[],
   opt?: GqlOpt,
-) {
+): Promise<number> {
+  if (ids.length === 0) {
+    return 0;
+  }
   const data: {
     revertByIdsSeo: Mutation["revertByIdsSeo"];
   } = await mutation({
@@ -344,7 +396,10 @@ export async function revertByIds(
 export async function forceDeleteByIds(
   ids: SeoId[],
   opt?: GqlOpt,
-) {
+): Promise<number> {
+  if (ids.length === 0) {
+    return 0;
+  }
   const data: {
     forceDeleteByIdsSeo: Mutation["forceDeleteByIdsSeo"];
   } = await mutation({
